simplify simple processor concurrency test
